+++ application/configuration/cfg.py
@@ -11,6 +11,7 @@ def get_bool(var, default):
 
 default_config = {
     'SECRET_KEY': environ.get('SECRET_KEY', 'secret_for_test'),
+    'ACTIVATION_ACCOUNT_TOKEN_MAX_AGE': get_int('ACTIVATION_ACCOUNT_TOKEN_MAX_AGE', 86400),
     'BCRYPT_DESIRED_KEY_BYTES': get_int('BCRYPT_DESIRED_KEY_BYTES', 32),
     'BCRYPT_ROUNDS': get_int('BCRYPT_DESIRED_KEY_BYTES', 100),
     'DATABASE_URL': environ.get('DATABASE_URL', 'sqlite:////tmp/:dashboard-plus:'),