+++ application/core/port/create_account_activation_token_port.py
@@ -1,0 +1,7 @@
+import abc
+
+
+class CreateAccountActivationTokenPort(metaclass=abc.ABCMeta):
+    @abc.abstractmethod
+    def create_account_activation_token(self, payload):
+        raise NotImplementedError('users must define create_account_activation_token to use this base class')