+++ application/core/port/create_safe_timed_token_port.py
@@ -1,7 +1,0 @@
-import abc
-
-
-class CreateSafeTimedTokenPort(metaclass=abc.ABCMeta):
-    @abc.abstractmethod
-    def create_safe_time_token(self, payload):
-        raise NotImplementedError('users must define create_safe_time_token to use this base class')