+++ application/core/port/decode_safe_time_token_port.py
@@ -1,0 +1,7 @@
+import abc
+
+
+class DecodeActivationAccountTokenPort(metaclass=abc.ABCMeta):
+    @abc.abstractmethod
+    def decode_activation_account_token(self, token: str) -> object:
+        raise NotImplementedError('users must define decode_activation_account_token to use this base class')