+++ tests/e2e/test_register_account_happy_path.py
@@ -1,6 +1,6 @@
 from application.configuration.ioc_usecase import IOCUseCase
 from application.core.usecase.usecase_input import UseCaseInput
-from application.core.usecase.usecase_output import UsecaseStatusEnum, UseCaseMessageEnum
+from application.core.usecase.usecase_output import UseCaseStatusEnum, UseCaseMessageEnum
 from tests.base_tests import E2ETest
 
 
@@ -18,7 +18,7 @@ class TestRegisterAccountHappyPath(E2ETest):
         usecase_output = usecase.handle(usecase_input)
 
         # Then
-        assert usecase_output.status == UsecaseStatusEnum.success
+        assert usecase_output.status == UseCaseStatusEnum.success
         assert usecase_output.message == UseCaseMessageEnum.account_registered
         assert usecase_output.content['account_id'] == 1
         assert usecase_output.content['activation_token']