+++ tests/integration/application/providers/data/test_account_database_data_provider.py
@@ -31,3 +31,39 @@ class TestAccountDatabaseDataProvider(IntegrationTest):
         with pytest.raises(EntityAlreadyExistsException):
             self.provider.insert(account)
             self.provider.insert(account)
+
+    def test_does_account_exist_should_return_true_when_account_exist(self):
+        # Given
+        account = AccountMapper(email='test@test.com')
+        self.db.session.add(account)
+        self.db.session.commit()
+
+        # When
+        result = self.provider.does_account_exist('test@test.com')
+
+        # Then
+        assert result is True
+
+    def test_does_account_exist_should_return_false_when_account_do_no_exist(self):
+        # Given
+        account = AccountMapper(email='test@test.com')
+        self.db.session.add(account)
+        self.db.session.commit()
+
+        # When
+        result = self.provider.does_account_exist('notfound@test.com')
+
+        # Then
+        assert result is False
+
+    def test_update_email_confirmed(self):
+        # Given
+        account = AccountMapper(email='test@test.com', email_confirmed=False)
+        self.db.session.add(account)
+        self.db.session.commit()
+
+        # When
+        self.provider.update_email_confirmed('test@test.com', True)
+
+        # Then
+        assert account.email_confirmed is True