+++ tests/unit/application/core/usecase/steps/test_create_account_activation_token_step.py
@@ -1,6 +1,6 @@
 from unittest import mock
 
-from application.core.port.create_safe_timed_token_port import CreateSafeTimedTokenPort
+from application.core.port.create_account_activation_token_port import CreateAccountActivationTokenPort
 from application.core.usecase.steps.create_account_activation_token_step import (
     CreateAccountActivationTokenStep
 )
@@ -9,11 +9,11 @@ from application.core.usecase.steps.create_account_activation_token_step import 
 class TestCreateAccountActivationTokenStep:
     def test_should_call_create_safe_time_token_on_the_provider(self):
         # Given
-        mock_token_provider = mock.create_autospec(CreateSafeTimedTokenPort)
+        mock_token_provider = mock.create_autospec(CreateAccountActivationTokenPort)
         step = CreateAccountActivationTokenStep(mock_token_provider)
 
         # When
         step.execute('data')
 
         # Then
-        mock_token_provider.create_safe_time_token.assert_called_with('data')
+        mock_token_provider.create_account_activation_token.assert_called_with('data')